Re Refactoring
